backend/requirements.lock.txt
aiofiles==23.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.12.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
exceptiongroup==1.3.0
faiss-cpu==1.7.4
fastapi==0.110.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.23.2
idna==3.10
intel-openmp==2021.4.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.325
langsmith==0.0.92
Markdown==3.8.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pydantic==2.7.2
pydantic_core==2.18.3
PyMuPDF==1.26.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.36.3
sympy==1.14.0
tbb==2021.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.67.1
transformers==4.43.1
typing-inspect==0.9.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.21.1
yarl==1.20.1
